Errors in code corrected, first unit and integration tests done

+++ src/main/java/com/projectbaby/Project/baby/repository/ActivityRepository.java
@@ -4,6 +4,6 @@ import com.projectbaby.Project.baby.model.entity.activity.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface ActivityRepository extends JpaRepository <Activity, Long> {
+public interface ActivityRepository extends JpaRepository <Activity, Integer> {
 
 }
