+++ src/main/java/com/projectbaby/Project/baby/repository/BabyRepository.java
@@ -4,6 +4,6 @@ import com.projectbaby.Project.baby.model.entity.Baby;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface BabyRepository extends JpaRepository<Baby, Long> {
+public interface BabyRepository extends JpaRepository<Baby, Integer> {
 
 }
